haup/requirements.txt
# HAUP v3.0 - Hybrid Adaptive Unified Pipeline
# Requirements for PostgreSQL/Neon Edition with Reverse Extraction & RAG

# Core Dependencies
python-dotenv>=1.0.0
psycopg2-binary>=2.9.9
chromadb>=0.4.22
sentence-transformers>=2.2.2
torch>=2.0.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0

# Terminal UI & Progress
rich>=13.7.0
psutil>=5.9.0

# RAG Engine - LLM Clients (Optional - choose based on your provider)
openai>=1.12.0              # For OpenAI GPT models
anthropic>=0.18.0           # For Anthropic Claude models
# Ollama uses built-in urllib, no extra package needed

# Excel Support (Optional - for Excel output)
openpyxl>=3.1.0

# CDC Support (Optional - for real-time sync)
kafka-python>=2.0.2

# Graph Support (Optional - for knowledge graphs)
# networkx>=3.0
# neo4j>=5.0.0

# Development & Testing (Optional)
# pytest>=7.4.0
# black>=23.0.0
# flake8>=6.0.0
